+++ Source/swiftlint-dev/SwiftLintDev.swift
@@ -1,0 +1,12 @@
+import ArgumentParser
+
+@main
+struct SwiftLintDev: AsyncParsableCommand {
+    static let configuration = CommandConfiguration(
+        commandName: "swiftlint-dev",
+        abstract: "A tool to help develop SwiftLint.",
+        subcommands: [
+            RuleTemplate.self,
+        ]
+    )
+}